+++ packages/core/src/main.ts
@@ -3,6 +3,7 @@ export * from "./components";
 export * from "./hooks";
 export * from "./constants";
 export * from "./styles";
+export * from "./router";
 
 import "./styles/global.css";