+++ packages/apps/browser/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @prozilla-os/browser
 
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @prozilla-os/core@1.3.2
+
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @prozilla-os/core@1.3.1
+
 ## 1.1.5
 
 ### Patch Changes