+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prozilla-os/core
 
+## 1.3.2
+
+### Patch Changes
+
+- Added vite plugin for site staging
+
+## 1.3.1
+
+### Patch Changes
+
+- Added ProzillaOS router
+
 ## 1.3.0
 
 ### Minor Changes