+++ packages/dev-tools/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @prozilla-os/dev-tools
 
+## 1.0.6
+
+### Patch Changes
+
+- Converted stageSitePlugin to rollup plugin
+
+## 1.0.5
+
+### Patch Changes
+
+- Fixed type declarations
+
+## 1.0.4
+
+### Patch Changes
+
+- Added missing exports
+
+## 1.0.3
+
+### Patch Changes
+
+- Added vite plugin for site staging
+- Updated dependencies
+  - @prozilla-os/core@1.3.2
+
+## 1.0.2
+
+### Patch Changes
+
+- Added ProzillaOS router
+- Updated dependencies
+  - @prozilla-os/core@1.3.1
+
 ## 1.0.1
 
 ### Patch Changes